Dynamically allocate memory for the inotify event buffer
By using malloc(3) depending on the number of files we have, the memory
footprint can be reduced.

Before this change, the buffer size was 32 * sizeof(struct inotify_event)
which was an overhead when just tailing some few files.

Now the memory is allocated depending on the number of files to tail. I
know of no occassion where I received more than one event at a time per
file, so the '2 *' can probably be dropped too.