inotail: Support reopening rotated files
We changed the behaviour of the "follow" option so that it detects that
file has been moved or deleted it keeps monitoring for the file instead
of ignoring it.  This is useful when tailing syslog circular logs, where
syslog periodically moves the log to another filename and recreates the
log file.  Without the patch inotail would issue a "file has been moved"
or a "file has been deleted message" and stop tracking the file. Now it
will close the file and refresh the inotify +descriptor. When a
notification arrives with a CREATE or MODIFY event, it will reopen the
file and set the size to zero.

We are using the patched program in an embedded ARM system with
satisfactory results.